@@ -10,11 +10,11 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.6.3</version>
+		<version>3.3.1</version>
 	</parent>
 
 	<properties>
-		<java.version>16</java.version>
+		<java.version>21</java.version>
 		<docker.image.prefix>fpw</docker.image.prefix>
 		<docker.image.name>frozen-particle-waves</docker.image.name>
 	</properties>
@@ -51,13 +51,6 @@
 		</dependency>
 
 		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-configuration-processor</artifactId>
-			<optional>true</optional>
-		</dependency>
-
-
-		<dependency>
 			<groupId>org.springframework</groupId>
 			<artifactId>spring-websocket</artifactId>
 		</dependency>
@@ -70,7 +63,6 @@
 		<dependency>
 			<groupId>org.projectlombok</groupId>
 			<artifactId>lombok</artifactId>
-			<version>1.18.22</version>
 			<scope>provided</scope>
 		</dependency>
 
