@@ -2,13 +2,13 @@ package fpw.security;
 
 import java.io.IOException;
 
-import javax.servlet.ServletException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
+
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 
 @Component
 public class RESTAuthenticationEntryPoint implements AuthenticationEntryPoint {
@@ -19,4 +19,5 @@ public class RESTAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
 		response.sendError(HttpServletResponse.SC_UNAUTHORIZED);
 	}
+
 }
